Add files via upload
This is a simple Chrome extension that allows users to download videos from YouTube and YouTube shorts with the click of a button. The extension is built using HTML, CSS, and JavaScript.

Installation
To install the extension, simply download the code from this GitHub repository, unzip the file, and follow these steps:

Open Google Chrome and navigate to chrome://extensions
Enable Developer Mode by toggling the switch in the top right corner of the page
Click the "Load unpacked" button in the top left corner of the page
Select the unzipped folder containing the extension code
Usage
Once the extension is installed, simply navigate to a YouTube video or YouTube short and click the "Download" button that appears below the video player. You can select the desired video format from the dropdown menu before clicking the download button.

Notes
This is a very basic extension and may not work for all YouTube videos. It was built as a learning exercise and is not intended for commercial use. Use at your own risk.